added catch for avoiding null pointer exception when deleting a role

+++ src/DiscordBot.java
@@ -344,6 +344,7 @@ public class DiscordBot {
 	 */
 	public static void deleteRole(IRole role)
 	{
+		if(role==null) return;
 		RequestBuffer.request(() -> {
 			role.delete();
 		});
